+++ src/controllers/api/account.ts
@@ -35,6 +35,11 @@ import { LocalStore } from '../../database/cache/store.js';
 import { BootStrapAccountResponse } from '../../types/account.js';
 import { RoleEntity } from '../../database/entities/role.entity.js';
 import { MailchimpService } from '../../helpers/mailchimp.js';
+import { IdentifierService } from '../../services/api/identifier.js';
+import { Credentials } from '../../services/api/credentials.js';
+import { ResourceService } from '../../services/api/resource.js';
+import { CredentialCategory } from '../../types/credential.js';
+import { Like } from 'typeorm';
 
 dotenv.config();
 
@@ -571,6 +576,79 @@ export class AccountController {
 		} catch (error) {
 			return response.status(StatusCodes.INTERNAL_SERVER_ERROR).json({
 				error: `Internal Error: ${(error as Error)?.message || error}`,
+			});
+		}
+	}
+
+	/**
+	 * @openapi
+	 *
+	 * /account/analytics:
+	 *   get:
+	 *     tags: [Account]
+	 *     summary: Fetch Account Analytics.
+	 *     description: This endpoint returns analytics data for the account usage.
+	 *     parameters:
+	 *       - in: query
+	 *         name: network
+	 *         description: Filter DID by the network published.
+	 *         schema:
+	 *           type: string
+	 *           enum:
+	 *             - mainnet
+	 *             - testnet
+	 *     responses:
+	 *       200:
+	 *         description: The request was successful.
+	 *         content:
+	 *           application/json:
+	 *             schema:
+	 *               $ref: '#/components/schemas/AccountAnalyticsResponse'
+	 *       400:
+	 *         $ref: '#/components/schemas/InvalidRequest'
+	 *       401:
+	 *         $ref: '#/components/schemas/UnauthorizedError'
+	 *       500:
+	 *         $ref: '#/components/schemas/InternalError'
+	 */
+	public async getAnalytics(request: Request, response: Response) {
+		const { network } = request.query;
+		try {
+			const [dids, credentials, accreditations, resources] = await Promise.all([
+				IdentifierService.instance.count(response.locals.customer, {
+					...(network && {
+						provider: `did:cheqd:${network}`,
+					}),
+				}),
+				Credentials.instance.count(response.locals.customer, {
+					category: CredentialCategory.CREDENTIAL,
+					deprecated: false,
+					...(network && {
+						issuerId: Like(`did:cheqd:${network}%`),
+					}),
+				}),
+				Credentials.instance.count(response.locals.customer, {
+					category: CredentialCategory.ACCREDITATION,
+					deprecated: false,
+					...(network && {
+						issuerId: Like(`did:cheqd:${network}%`),
+					}),
+				}),
+				ResourceService.instance.count(response.locals.customer, {
+					...(network && {
+						identifier: Like(`did:cheqd:${network}%`),
+					}),
+				}),
+			]);
+			return response.status(StatusCodes.OK).json({
+				dids,
+				credentials,
+				accreditations,
+				resources,
+			});
+		} catch (error) {
+			return response.status(StatusCodes.INTERNAL_SERVER_ERROR).json({
+				error: `Internal error: ${(error as Error)?.message || error}`,
 			});
 		}
 	}